moved findClauseInLibrary to a shared location for helpers to navigate ELM
